+++ apps/expo-go/ios/Exponent/Versioned/Core/EXVersionManagerObjC.mm
@@ -117,11 +117,11 @@ RCT_EXTERN void EXRegisterScopedModule(Class, ...);
     [[RCTPackagerConnection sharedPackagerConnection] reconnect:packagerServerHostPort];
     RCTInspectorPackagerConnection *inspectorPackagerConnection = [RCTInspectorDevServerHelper connectWithBundleURL:bundleURL];
 
-    NSDictionary<NSString *, id> *buildProps = [self.manifest getPluginPropertiesWithPackageName:@"expo-build-properties"];
-    NSNumber *enableNetworkInterceptor = [[buildProps objectForKey:@"ios"] objectForKey:@"unstable_networkInspector"];
-    if (enableNetworkInterceptor == nil || [enableNetworkInterceptor boolValue] != NO) {
-      self.networkInterceptor = [[EXVersionedNetworkInterceptor alloc] initWithRCTInspectorPackagerConnection:inspectorPackagerConnection];
-    }
+//    NSDictionary<NSString *, id> *buildProps = [self.manifest getPluginPropertiesWithPackageName:@"expo-build-properties"];
+//    NSNumber *enableNetworkInterceptor = [[buildProps objectForKey:@"ios"] objectForKey:@"unstable_networkInspector"];
+//    if (enableNetworkInterceptor == nil || [enableNetworkInterceptor boolValue] != NO) {
+//      self.networkInterceptor = [[EXVersionedNetworkInterceptor alloc] initWithRCTInspectorPackagerConnection:inspectorPackagerConnection];
+//    }
   }
 }