+++ packages/expo-audio/ios/AudioModule.swift
@@ -165,10 +165,8 @@ public class AudioModule: Module {
     // swiftlint:disable:next closure_body_length
     Class(AudioRecorder.self) {
       Constructor { (options: RecordingOptions) -> AudioRecorder in
-        guard let cachesDir = appContext?.fileSystem?.cachesDirectory, let directory = URL(string: cachesDir) else {
-          throw Exceptions.AppContextLost()
-        }
-        let avRecorder = AudioUtils.createRecorder(directory: directory, with: options)
+        let recordingDir = try recordingDirectory()
+        let avRecorder = AudioUtils.createRecorder(directory: recordingDir, with: options)
         let recorder = AudioRecorder(avRecorder)
         AudioComponentRegistry.shared.add(recorder)
 
@@ -191,26 +189,23 @@ public class AudioModule: Module {
         recorder.uri
       }
 
+      AsyncFunction("prepareToRecordAsync") { (recorder, options: RecordingOptions?) in
+        recorder.prepare(options: options)
+      }
+
       Function("record") { (recorder: AudioRecorder) -> [String: Any] in
         try checkPermissions()
-        recorder.ref.record()
-        recorder.startTimestamp = Int(recorder.deviceCurrentTime)
-        return recorder.getRecordingStatus()
+        return recorder.startRecording()
       }
 
       Function("pause") { recorder in
         try checkPermissions()
-        recorder.ref.pause()
-        let current = recorder.deviceCurrentTime
-        recorder.previousRecordingDuration += (current - recorder.startTimestamp)
-        recorder.startTimestamp = 0
+        recorder.pauseRecording()
       }
 
-      Function("stop") { recorder in
+      AsyncFunction("stop") { recorder in
         try checkPermissions()
-        recorder.ref.stop()
-        recorder.startTimestamp = 0
-        recorder.previousRecordingDuration = 0
+        recorder.stopRecording()
       }
 
       Function("getStatus") { recorder -> [String: Any] in
@@ -241,6 +236,13 @@ public class AudioModule: Module {
     }
   }
 
+  private func recordingDirectory() throws -> URL {
+    guard let cachesDir = appContext?.fileSystem?.cachesDirectory, let directory = URL(string: cachesDir) else {
+      throw Exceptions.AppContextLost()
+    }
+    return directory
+  }
+
   private func setIsAudioActive(_ isActive: Bool) throws {
     if !isActive {
       for player in AudioComponentRegistry.shared.players.values {
@@ -265,6 +267,7 @@ public class AudioModule: Module {
       AudioComponentRegistry.shared.recorders.values.forEach { recorder in
         if recorder.isRecording {
           recorder.ref.stop()
+          recorder.allowsRecording = false
         }
       }
     }