+++ packages/expo-audio/src/ExpoAudio.ts
@@ -15,22 +15,24 @@ import { AudioPlayer, AudioRecorder } from './AudioModule.types';
 import { createRecordingOptions } from './utils/options';
 import { resolveSource } from './utils/resolveSource';
 
+export const PLAYBACK_STATUS_UPDATE = 'playbackStatusUpdate';
+export const AUDIO_SAMPLE_UPDATE = 'audioSampleUpdate';
+export const RECORDING_STATUS_UPDATE = 'recordingStatusUpdate';
+
 export function useAudioPlayer(
   source: AudioSource | string | number | null = null,
   updateInterval: number = 500
 ): AudioPlayer {
   const parsedSource = resolveSource(source);
-  const player = useReleasingSharedObject(
+  return useReleasingSharedObject(
     () => new AudioModule.AudioPlayer(parsedSource, updateInterval),
     [JSON.stringify(parsedSource)]
   );
-
-  return player;
 }
 
 export function useAudioPlayerStatus(player: AudioPlayer): AudioStatus {
   const currentStatus = useMemo(() => player.currentStatus, [player.id]);
-  return useEvent(player, 'onPlaybackStatusUpdate', currentStatus);
+  return useEvent(player, PLAYBACK_STATUS_UPDATE, currentStatus);
 }
 
 export function useAudioSampleListener(
@@ -42,7 +44,7 @@ export function useAudioSampleListener(
     if (!player.isAudioSamplingSupported) {
       return;
     }
-    const subscription = player.addListener('onAudioSampleUpdate', listener);
+    const subscription = player.addListener(AUDIO_SAMPLE_UPDATE, listener);
     return () => {
       subscription.remove();
     };
@@ -59,7 +61,7 @@ export function useAudioRecorder(
   }, [JSON.stringify(platformOptions)]);
 
   useEffect(() => {
-    const subscription = recorder.addListener('onRecordingStatusUpdate', (status) => {
+    const subscription = recorder.addListener(RECORDING_STATUS_UPDATE, (status) => {
       statusListener?.(status);
     });
     return () => subscription.remove();