+++ packages/create-expo/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+## 3.0.1 — 2024-10-22
+
+### 💡 Others
+
 - Update `tar` dependency. ([#29663](https://github.com/expo/expo/pull/29663) by [@Simek](https://github.com/Simek))
 - Use `npx` for npm examples. ([#31012](https://github.com/expo/expo/pull/31012) by [](@kadikraman)(https://github.com/kadikraman))
 
@@ -23,6 +27,7 @@
 ### 🎉 New features
 
 ### 🐛 Bug fixes
+
 - Update list of available templates ([#29955](https://github.com/expo/expo/pull/29955) by [@kadikraman](https://github.com/kadikraman))
 
 ### 💡 Others