+++ packages/expo-dev-client/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unpublished
 
+### 🛠 Breaking changes
+
+### 🎉 New features
+
+### 🐛 Bug fixes
+
+### 💡 Others
+
+## 5.0.0-preview.0 — 2024-10-22
+
 - Add `addGeneratedScheme` config plugin option to disable automatic adding of dev client app scheme. ([#31147](https://github.com/expo/expo/pull/31147) by [@keith-kurak](https://github.com/keith-kurak))
 
 ### 🛠 Breaking changes